+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `reconfigure`:
   - Parameters can be now used within conditions.
   - Implementation of `__contains__` to support `if ... in P`.
+  - `update()` now takes optional argument `only_existing` (def. `False`) to only update existing parameters.
 
 ## 0.5.1 - 2022-03-23
 ### Added